refactor(gateway): isolate session describe handler
